requirements.txt
biopython>=1.78
matplotlib>=3.3.0
